♻️ : shared/ui Table 분리
