+++ src/shared/ui/table/index.ts
@@ -1,0 +1,6 @@
+export { Table } from "./Table"
+export { TableHeader } from "./TableHeader"
+export { TableBody } from "./TableBody"
+export { TableRow } from "./TableRow"
+export { TableHead } from "./TableHead"
+export { TableCell } from "./TableCell"
